+++ pyhacrf/pyhacrf.py
@@ -104,7 +104,7 @@ class Hacrf(object):
                 ll += dll
                 gradient += dgradient
 
-            parameters_without_bias = np.array(parameters)  # exclude the bias parameters from being regularized
+            parameters_without_bias = np.array(parameters, dtype='float64')  # exclude the bias parameters from being regularized
             parameters_without_bias[0] = 0
             ll -= self.l2_regularization * np.dot(parameters_without_bias.T, parameters_without_bias)
             gradient = gradient.flatten() - 2.0 * self.l2_regularization * parameters_without_bias
@@ -259,7 +259,7 @@ class _Model(object):
         beta = self._backward(x_dot_parameters)
         classes_to_ints = {k: i for i, k in enumerate(set(self.states_to_classes.values()))}
         states_to_classes = np.array([classes_to_ints[self.states_to_classes[state]]
-                                      for state in range(max(self.states_to_classes.keys()) + 1)])
+                                      for state in range(max(self.states_to_classes.keys()) + 1)], dtype='int64')
         if not isinstance(self.sparse_x, str):
             ll, deriv = gradient_sparse(alpha, beta, parameters, states_to_classes,
                                         self.sparse_x[0], self.sparse_x[1], classes_to_ints[self.y],
